feat: add scoreChange component function
This function will be used by onClick listeners on the logos so that when a logo is clicked it will either increase the currentScore, or increase bestScore and reset currentScore.

feat: add heading & paragraph to component return

The heading and paragraph detail to the user what this application is and how it is played.

feat: add score paragraphs

These paragraphs show the dynamic values of currentScore and bestScore to the user while they play.

feat: import Card

Card is a component that will be used to build an element for each mlbTeams array element, aka an element for each mlb team and use their team abbreviation and logo in the element.

feat: add instances of Card

Using map we efficiently build an instance of Card for each team and pass their abbreviation and logo to Card. We also pass an onClick function that increments the appropriate score and calls shuffle to randomize the team elements order.