feat: create Card component file
This file will be used to build Card components and will be imported into App.

feat: add Card props

App will pass to Card instances values for 3 separate props, two coming from a specific team object from the fetched data, the other a series of actions to perform when the element is clicked on.

feat: add section return to Card component

Card will build an element that displays a teams logo in its background, and list that teams abbreviation for users to also read. When a user clicks on this Card's returned element it will invoke the onClick function that increments a specific score state and then causes the teams to be shuffled and rerendered.