style: change height to 100px
style: change width to min and 15vw

This scales for mobile devices as before the elements were too wide to fit multiple on a horizontal line but now scale.

style: add background-position of center

This wasn't needed before but now with responsive width it is needed.

style: change background-size to contain

With the responsive changes 'cover' did not look good because each logo was zoomed in too much, contain is perfect here.

style: add color to text

Using white text helps the abbreviations stick out.

style: change text-align to start

With the changes to dynamic sizing when the Cards where scaled to a desktop size the background logo images were behind the text and made some hard to read so changed it to start align, and that barely works for mobile readability but it does.

style: add padding-left to help with abbreviations

Adding this helps abbreviations stay away from border and increase readability.