fix: move api fetch to a useEffect
This was something I was hung on, I couldn't think of how to use a fetch call in a useEffect with async and my brain kept saying "but the component is going to try to render before the effect is invoked to get the data" well no it won't because if we use map on mlbTeams when it's empty, it won't do anything. So by using an empty dependency array this effect only runs on mount, which is good because I don't want to piss off ESPN by constantly fetching on every user turn. A cleanup function was not used because this effect is only invoking when the component mounts initially, which is when the page is loaded or refreshed in the browser.

fix: move mlbTeams into state

mlbTeams changes over time, therefore it should be in state, something I read and it clicked "I've read this before" oh well I need to learn somehow.

feat: add shuffle invocation in useEffect

I wanted to shuffle the teams for the first appearance of the team logos to the user but I couldn't figure out how, it's because everything I tried previously was changing the initial mlbTeams state, which is empty before the useEffect fetch changes it and causes a rerender. So now the data is fetched and then shuffle takes that data and randomizes it and then returns it, which will be used by the mlbTeams state setting function to save that order.

fix: wrap shuffle with set function

We save a line here by wrapping the return of shuffle inside setMlbTeams